+++ acapy_agent/anoncreds/models/issuer_cred_rev_record.py
