+++ acapy_agent/anoncreds/revocation/setup.py
@@ -5,11 +5,11 @@ from abc import ABC, abstractmethod
 
 from acapy_agent.protocols.endorse_transaction.v1_0.util import is_author_role
 
-from ..anoncreds.revocation import AnonCredsRevocation, AnonCredsRevocationError
-from ..core.event_bus import EventBus
-from ..core.profile import Profile
-from ..revocation.util import notify_revocation_published_event
-from .events import (
+from .revocation import AnonCredsRevocation, AnonCredsRevocationError
+from ...core.event_bus import EventBus
+from ...core.profile import Profile
+from ...revocation.util import notify_revocation_published_event
+from ..events import (
     CRED_DEF_FINISHED_PATTERN,
     REV_LIST_FINISHED_PATTERN,
     REV_REG_DEF_FINISHED_PATTERN,
